% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_connector.R
\name{ssh_authenticator}
\alias{ssh_authenticator}
\title{Authentication to SSH via Shiny module (within teal) or manually}
\usage{
ssh_authenticator(id = "cred", default_host = NULL)
}
\arguments{
\item{id}{(\code{character(1)}) An ID string that corresponds with the ID used to call the module's UI function.
Defaults to \code{"cred"}.}

\item{default_host}{(\code{character(1)}) optional host definition that forces this
option on the user. If it is not defined, then user can define themselves.}
}
\value{
A list with a session object from \code{\link[ssh:ssh]{ssh::ssh_connect()}}.
}
\description{
It should be used as part of a \code{\link[teal:teal_data_module]{teal::teal_data_module()}} server to
perform SSH authentication in a teal application.

This also allows to authenticate with the reproducible generated by a teal application.
}
\examples{
\dontrun{
ssh_authenticator("demo")
}
}
