% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.connectors.ssh.R
\docType{package}
\name{teal.connectors.ssh}
\alias{teal.connectors.ssh-package}
\alias{teal.connectors.ssh}
\title{teal.modules.ssh: functions and class constructors to connect to SSH data}
\description{
The functions in this package is meant to construct R6 data and data
connector classes from the teal package
}
\author{
\strong{Maintainer}: Ande Verissimo \email{andre.verissimo@roche.com} (\href{https://orcid.org/0000-0002-2212-339X}{ORCID})

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
