% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_connector.R
\name{ssh_authenticator}
\alias{ssh_authenticator}
\title{Authentication to SSH via Shiny module (within teal) or manually}
\usage{
ssh_authenticator(host = NULL)
}
\arguments{
\item{host}{(\code{character(1)}) Optional host address for SSH connection.}
}
\value{
A list with a session object from \code{\link[ssh:ssh]{ssh::ssh_connect()}}.
}
\description{
It should be used as part of a \code{\link[teal:teal_data_module]{teal::teal_data_module()}} server to
perform SSH authentication in a teal application.

This also allows to authenticate with the reproducible generated by a teal application.
}
\examples{
\dontrun{
ssh_authenticator()
}
}
